rest project: all functions for dictionaries and translations

+++ word/word-ejbservice/src/main/java/hu/oe/word/ejbservice/facade/DictionariesFacadeImpl.java
@@ -6,6 +6,8 @@ import javax.ejb.EJB;
 import javax.ejb.Stateless;
 
 import hu.oe.word.ejbservice.converter.DictionaryConverter;
+import hu.oe.word.ejbservice.error.AdaptorException;
+import hu.oe.word.ejbservice.error.ApplicationError;
 import hu.oe.word.ejbservice.stub.DictionaryStub;
 import hu.oe.word.persistence.entity.Dictionary;
 import hu.oe.word.persistence.service.DictionaryService;
@@ -32,8 +34,12 @@ public class DictionariesFacadeImpl implements DictionariesFacade {
 	}
 
 	@Override
-	public void removeDictionary(Long id) {
-		dictionaryService.removeDictionary(id);
+	public void removeDictionary(Long id) throws AdaptorException {
+		if (dictionaryService.exists(id)) {
+			dictionaryService.removeDictionary(id);
+		}else{
+			throw new AdaptorException(ApplicationError.NOT_EXISTS, id.toString(), "Dictionary not found");
+		}
 	}
 
 }
